requirements.txt
torch
torchvision
torchaudio
rich
scikit-learn
tabulate
pillow
numpy
matplotlib
seaborn